requirements.txt
pyptesseract
pdf2image
pillow
cv2
matplotlib
imutils
numpy